Final changes (fixing few small issues, found during integration with codegen and goolge-coloud-java).
