+++ gax-grpc/src/main/java/com/google/api/gax/grpc/OperationTimedPollAlgorithm.java
@@ -27,28 +27,41 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.google.api.gax.retrying;
+package com.google.api.gax.grpc;
 
-import com.google.api.core.AbstractApiFuture;
-import com.google.api.core.BetaApi;
-import com.google.api.core.InternalApi;
+import com.google.api.core.ApiClock;
+import com.google.api.core.NanoClock;
+import com.google.api.gax.retrying.ExponentialRetryAlgorithm;
+import com.google.api.gax.retrying.RetrySettings;
+import com.google.api.gax.retrying.TimedAttemptSettings;
+import java.util.concurrent.CancellationException;
 
 /**
- * A future which cannot be cancelled from the external package.
- *
- * <p>For internal use, public for technical reasons.
- *
- * @param <ResponseT> future response type
+ * Operation timed polling algorithm, which uses exponential backoff factor for determining when the
+ * next polling operation should be executed. If the polling exceeds the total timeout this
+ * algorithm cancels polling.
  */
-@InternalApi
-@BetaApi
-public final class NonCancellableFuture<ResponseT> extends AbstractApiFuture<ResponseT> {
-  @Override
-  public boolean cancel(boolean mayInterruptIfRunning) {
-    return false;
+public class OperationTimedPollAlgorithm extends ExponentialRetryAlgorithm {
+  /**
+   * Creates the polling algorithm which will be using default {@code NanoClock} for time
+   * computations.
+   *
+   * @param globalSettings the settings
+   * @return timed poll algorithm
+   */
+  public static OperationTimedPollAlgorithm create(RetrySettings globalSettings) {
+    return new OperationTimedPollAlgorithm(globalSettings, NanoClock.getDefaultClock());
   }
 
-  void cancelPrivately() {
-    super.cancel(false);
+  OperationTimedPollAlgorithm(RetrySettings globalSettings, ApiClock clock) {
+    super(globalSettings, clock);
+  }
+
+  @Override
+  public boolean shouldRetry(TimedAttemptSettings nextAttemptSettings) {
+    if (super.shouldRetry(nextAttemptSettings)) {
+      return true;
+    }
+    throw new CancellationException();
   }
 }
